Cast enum constants to 32-bit integers
When generating C++ and C# code, gbc now emits casts to allow uint32
enum constant values to be used without errors or warnings.

* Use explicit static_cast<int32_t> for C++ enum constants to avoid the
  Clang warning -Wc++11-narrowing when using uint32 enum constants.
* Use unchecked((int)value) for C# enum constants to avoid a compiler
  error.

+++ compiler/src/Language/Bond/Codegen/Cpp/Util.hs
@@ -139,7 +139,6 @@ enumDefinition Enum {..} = [lt|enum #{declName}
         };|]
   where
     constant Constant {..} = [lt|#{constantName}#{optional value constantValue}|]
-    value (-2147483648) = [lt| = -2147483647-1|]
-    value x = [lt| = #{x}|]
+    value (-2147483648) = [lt| = static_cast<int32_t>(-2147483647-1)|]
+    value x = [lt| = static_cast<int32_t>(#{x})|]
 enumDefinition _ = error "enumDefinition: impossible happened."
-
